# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Tenant PIM API
  version: 2.4.0
  description: |
    API para gestión de información de productos multi-tenant.

    ## Arquitectura Modular

    El sistema está organizado en módulos especializados:
    - **Health**: Estado del sistema
    - **Business Types**: Tipos de negocio
    - **Brands**: Gestión de marcas
    - **Products**: Gestión de productos y variantes
    - **Variants**: Gestión de variantes de productos (HITO 2.1)
    - **Categories**: Gestión de categorías
    - **Attributes**: Gestión de atributos
    - **Category-Attributes**: Relación categoría-atributo
    - **Global Catalog**: Catálogo global de productos con enrichment y quality score
    - **Global Catalog Public**: Endpoints públicos del catálogo global (sin auth)
    - **Product Requests**: Solicitudes de productos por tenants
    - **Quickstart**: Configuración rápida e importación desde global catalog
    - **Import**: Importación masiva con variantes (HITO 2.1)

    ## Actualizado v2.3.0

    - Global Catalog: CRUD completo, búsqueda por EAN, enrichment queue
    - Product Requests: Solicitudes de tenants con flujo approve/reject/fulfill
    - Quickstart: Import desde global catalog, import masivo por business type
    - Public endpoints: Búsqueda, sugerencias, facets sin autenticación
    - Variantes de productos: Product -> Variants con SKU único
    - Templates Quickstart: Configuración rápida por tipo de negocio
...
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para autenticación
    tenantId:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: ID del tenant para operaciones multi-tenant
    userRole:
      type: apiKey
      in: header
      name: X-User-Role
      description: |
        Rol del usuario para autorización:
        - `super_admin`: Acceso completo
        - `marketplace_admin`: Gestión de marketplace
        - `tenant_admin`: Administración del tenant
        - `tenant_user`: Usuario del tenant
...
paths:
...
  /products:
    get:
      summary: Listar productos con filtros
      tags: [Products]
      security:
        - BearerAuth: []
        - tenantId: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortDirParam'
        - name: name
          in: query
          description: Filtrar por nombre
          schema:
            type: string
        - name: sku
          in: query
          description: Filtrar por SKU
          schema:
            type: string
        - name: status
          in: query
          description: Filtrar por estado
          schema:
            type: string
            enum: ["active", "inactive", "discontinued"]
        - name: category_id
          in: query
          description: Filtrar por ID de categoría
          schema:
            type: string
            format: uuid
        - name: brand_id
          in: query
          description: Filtrar por ID de marca
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista paginada de productos
        '400':
          description: Parámetros inválidos
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor

    post:
      summary: Crear nuevo producto
      tags: [Products]
      security:
        - BearerAuth: []
        - tenantId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, sku]
              properties:
                name:
                  type: string
                  example: "Smartphone Galaxy S23"
                description:
                  type: string
                  example: "Smartphone de última generación"
                sku:
                  type: string
                  example: "SAM-GAL-S23-001"
                category_id:
                  type: string
                  format: uuid
                brand_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Producto creado exitosamente
        '400':
          description: Datos de entrada inválidos
        '409':
          description: SKU ya existe
        '500':
          description: Error interno del servidor
  /products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Obtener producto por ID
      tags: [Products]
      security:
        - BearerAuth: []
        - tenantId: []
      responses:
        '200':
          description: Detalles del producto
        '404':
          description: Producto no encontrado
        '500':
          description: Error interno del servidor

    put:
      summary: Actualizar producto
      tags: [Products]
      security:
        - BearerAuth: []
        - tenantId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                category_id:
                  type: string
                  format: uuid
                brand_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Producto actualizado exitosamente
        '400':
          description: Datos de entrada inválidos
        '404':
          description: Producto no encontrado
        '500':
          description: Error interno del servidor

    delete:
      summary: Eliminar producto
      tags: [Products]
      security:
        - BearerAuth: []
        - tenantId: []
      responses:
        '204':
          description: Producto eliminado
        '404':
          description: Producto no encontrado
        '500':
          description: Error interno del servidor
...
  /global-catalog/products:
    post:
      summary: Crear producto global
      tags: [Global Catalog]
      description: |
        Crea un producto en el catálogo global. Usado principalmente por scrapers
        y procesos de enrichment automatizado.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserRoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlobalProductCreateRequest'
      responses:
        '201':
          description: Producto global creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalProduct'
        '400':
          description: Datos de entrada inválidos
        '401':
          description: No autorizado
        '409':
          description: Producto con EAN duplicado
        '500':
          description: Error interno del servidor

    get:
      summary: Listar productos globales con filtros avanzados
      tags: [Global Catalog]
      description: |
        Lista productos del catálogo global con filtros avanzados.
        Soporta búsqueda por texto (ILIKE), filtros exactos y paginación.
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          description: Búsqueda por nombre (ILIKE)
          schema:
            type: string
        - name: description
          in: query
          description: Búsqueda por descripción (ILIKE)
          schema:
            type: string
        - name: brand
          in: query
          description: Filtrar por marca exacta
          schema:
            type: string
        - name: category
          in: query
          description: Filtrar por categoría exacta
          schema:
            type: string
        - name: business_type
          in: query
          description: Filtrar por tipo de negocio
          schema:
            type: string
        - name: source
          in: query
          description: Filtrar por fuente (scraper)
          schema:
            type: string
        - name: ean
          in: query
          description: Filtrar por código EAN
          schema:
            type: string
        - name: is_verified
          in: query
          description: Solo productos verificados
          schema:
            type: boolean
        - name: is_active
          in: query
          description: Solo productos activos
          schema:
            type: boolean
        - name: min_quality
          in: query
          description: Quality score mínimo (0-100)
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: max_quality
          in: query
          description: Quality score máximo (0-100)
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Lista paginada de productos globales
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalProductListResponse'
        '400':
          description: Parámetros de filtrado inválidos
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor
  /global-catalog/products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: ID del producto global

    get:
      summary: Obtener producto global por ID
      tags: [Global Catalog]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Detalles del producto global
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalProduct'
        '404':
          description: Producto no encontrado
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor

    put:
      summary: Actualizar producto global
      tags: [Global Catalog]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserRoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlobalProductUpdateRequest'
      responses:
        '200':
          description: Producto global actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalProduct'
        '400':
          description: Datos de entrada inválidos
        '404':
          description: Producto no encontrado
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor

    delete:
      summary: Eliminar producto global
      tags: [Global Catalog]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserRoleHeader'
      responses:
        '204':
          description: Producto global eliminado
        '404':
          description: Producto no encontrado
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor
  /global-catalog/products/search:
    get:
      summary: Búsqueda de productos por EAN
      tags: [Global Catalog]
      description: Busca productos en el catálogo global por código EAN.
      security:
        - BearerAuth: []
      parameters:
        - name: ean
          in: query
          required: true
          description: Código EAN del producto
          schema:
            type: string
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalProduct'
        '404':
          description: Producto no encontrado por EAN
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor
  /global-catalog/enrichment-queue:
    get:
      summary: Cola de enrichment
      tags: [Global Catalog]
      description: |
        Retorna productos del catálogo global con quality_score menor a 70,
        priorizados para enrichment (mejora de datos).
      security:
        - BearerAuth: []
      parameters:
        - name: business_type
          in: query
          description: Filtrar por tipo de negocio
          schema:
            type: string
        - name: limit
          in: query
          description: Cantidad máxima de resultados
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Offset para paginación
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Productos pendientes de enrichment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrichmentQueueResponse'
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor

  # ==========================================
  # PRODUCT REQUESTS ENDPOINTS (v2.3.0)
  # ==========================================
...
  /public/global-catalog/search:
    get:
      summary: Búsqueda pública por EAN
      tags: [Global Catalog Public]
      description: |
        Busca productos en el catálogo global por código EAN.
        No requiere autenticación.
      parameters:
        - name: ean
          in: query
          required: true
          description: Código EAN del producto
          schema:
            type: string
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalProduct'
        '404':
          description: Producto no encontrado
        '500':
          description: Error interno del servidor
...